+++ lib/src/visualization/extension.dart
@@ -1,5 +1,4 @@
 import 'dart:io';
-import 'dart:typed_data';
 
 import 'package:growth_standards/src/common/base.dart';
 import 'package:growth_standards/src/common/trajectory.dart';
@@ -12,17 +11,9 @@ extension GrowthChartResultVisualizationExtension on Result {
   String toSvg({GrowthChartConfig? config}) =>
       GrowthChartVisualizer.toSvg(this, config: config);
 
-  /// Render calculation result into PNG image bytes
-  Uint8List toPng({GrowthChartConfig? config}) =>
-      GrowthChartVisualizer.toPng(this, config: config);
-
   /// Save calculation result to an SVG file
   File saveSvg(String filePath, {GrowthChartConfig? config}) =>
       GrowthChartVisualizer.saveSvg(this, filePath, config: config);
-
-  /// Save calculation result to a PNG file
-  File savePng(String filePath, {GrowthChartConfig? config}) =>
-      GrowthChartVisualizer.savePng(this, filePath, config: config);
 }
 
 /// Extension on [List<Result>] for rendering multi-point growth trajectories
@@ -31,17 +22,9 @@ extension GrowthChartResultListVisualizationExtension on List<Result> {
   String toSvg({GrowthChartConfig? config}) =>
       GrowthChartVisualizer.toSvgList(this, config: config);
 
-  /// Render multiple calculation results (growth trajectory) into PNG image bytes
-  Uint8List toPng({GrowthChartConfig? config}) =>
-      GrowthChartVisualizer.toPngList(this, config: config);
-
   /// Save growth trajectory to an SVG file
   File saveSvg(String filePath, {GrowthChartConfig? config}) =>
       GrowthChartVisualizer.saveSvgList(this, filePath, config: config);
-
-  /// Save growth trajectory to a PNG file
-  File savePng(String filePath, {GrowthChartConfig? config}) =>
-      GrowthChartVisualizer.savePngList(this, filePath, config: config);
 }
 
 /// Extension on [VelocityBasedResult] for rendering growth velocity charts
@@ -50,17 +33,9 @@ extension VelocityBasedResultVisualizationExtension on VelocityBasedResult {
   String toSvg({GrowthChartConfig? config}) =>
       GrowthChartVisualizer.renderVelocitySvg(this, config: config);
 
-  /// Render growth velocity calculation into PNG image bytes
-  Uint8List toPng({GrowthChartConfig? config}) =>
-      GrowthChartVisualizer.renderVelocityPng(this, config: config);
-
   /// Save growth velocity calculation to an SVG file
   File saveSvg(String filePath, {GrowthChartConfig? config}) =>
       GrowthChartVisualizer.saveVelocitySvg(this, filePath, config: config);
-
-  /// Save growth velocity calculation to a PNG file
-  File savePng(String filePath, {GrowthChartConfig? config}) =>
-      GrowthChartVisualizer.saveVelocityPng(this, filePath, config: config);
 }
 
 /// Extension on [GrowthTrajectory] for direct chart rendering
@@ -68,14 +43,7 @@ extension GrowthTrajectoryVisualizationExtension on GrowthTrajectory {
   /// Render growth trajectory into an SVG vector graphic string
   String toSvg({GrowthChartConfig? config}) => results.toSvg(config: config);
 
-  /// Render growth trajectory into PNG image bytes
-  Uint8List toPng({GrowthChartConfig? config}) => results.toPng(config: config);
-
   /// Save growth trajectory to an SVG file
   File saveSvg(String filePath, {GrowthChartConfig? config}) =>
       results.saveSvg(filePath, config: config);
-
-  /// Save growth trajectory to a PNG file
-  File savePng(String filePath, {GrowthChartConfig? config}) =>
-      results.savePng(filePath, config: config);
 }